Remove-Duplication TestDelete class
